fixed block by smith
