+++ Source/JackFramework/ImgHandler/img_handler.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-import numpy as np
 import torch
 import torch.nn as nn
